refactor TypeBroadcast
